Change the way potion brew checking is done. Ignore physical interactions for now.
Because the old method of checking potions left much to be desired, I
found a way to use the NMS brewing methods to find the result of a
recipe. The "nobrew" permssion now uses only the data value of the
RESULTING potion, not the recipe. (e.g. noitem.nobrew.64)
I have also skipped the checking of physical interactions for now, since
they require a workaround to avoid causing massive lag.