Added armor check for PlayerJoinEvent
This will make sure that a player can still wear the armor they are
wearing when they join the server. This is so that a player can't keep
using armor they are no longer allowed to use. Armor will be put back
into the player's inventory or dropped on the ground if the inventory is
full.